+++ pyatlas_checks/pyatlas_checks/__init__.py
@@ -1,0 +1,4 @@
+"""
+A Python CLI for Atlas-Checks tools
+"""
+from pyatlas_checks.commands import *
